configurable TTY item : do not automatically scan port list
Previously when calling diag_l0_new() on an interface that creates a tty cfgi, tty_refresh() was automatically called. This can be very slow since it opens + closes every known device. It's still possible to find available ports using "set port ?", where the extra delay should not be a problem.