tests:  add test + fix for 14230 responses that start with 00 LEN:
(see GH issue #10). Previously, the L2 iso14230 code would ignore the first byte of received data if it was a 0x00.
This was because a fast init wake-up pattern (25ms break on TXD) indeed generates a spurious 0x00 byte, and would corrupt the following header. The TTY layer is not designed to distinguish between a break condition and regular data.
But this situation arises only in monitor mode (the L2 fastinit code already filters out the incoming 0x00 echo).
This "fix" now only drops the 0x00 byte if the connection was started in Monitor mode. The downside is now that in monitor mode, those problematic iso14230 frames are still problematic.
Examples of received data :
"00 83 10 FC 81 <cks> " received in monitor mode, the first 0x00 is invalid.
"01 7E <cks>"
"81 FC 10 3E <cks> : both typical, common iso14230 packets. No problems here.
"00 01 7E <cks>" : valid 14230 packet, but the 00 will be dropped in monitor mode.