L2 StartCommunications: do not allow multiple L2 conns to reuse the same L0-L2 link.
The only time where this would be of use would be to talk to multiple ECUs on the same link; this is cool but freediag's structure is just not meant for that. Some problems:
- how to protect the L0 and L2 data structures for multiple users
- how to sequence the data into L0 ? some kind of "worker thread" with a message queue would be required
- the CLI just can't use that feature

the cleanest way to do so would be to follow the structure of the J2534 API, with a thread handling L0 and L2, providing clean TX/RX message queues. This simplifies monitoring (just set the filters to enable all), periodic messages, connection reconfiguration, etc. Also, J2534.