diag_os: permit static allocation of mutexes, apply to only mutex
Allocation is removed from diag_os_newmtx and it is renamed to
diag_os_initmtx. diag_os_delmtx no longer frees.

The void typedef had to be corrected for static allocation of mutex
objects to be possible. (Note that no objects can be void in C, ie.
void is a null set.)