errors: split error generation and forwarding
Previously, if an error was 'sourced' at a certain level by calling
diag_iseterr() or diag_pseterr(), usually the caller also forwarded
the same error in the same way.
This produced a lot of output mostly consisting of "stacks" of
identical errors.
Now, diag_[i,p]seterr() is only used when an error is generated. If
only forwarding an error code to the caller, diag_[i,p]fwderr() can
be used which sends the message to the debugging sink instead of
stderr.

Note, currently the "debugging sink" is currently stderr (WIP), so
this commit doesn't change the visible output.

To make this change possible, a few minor tweaks were required
- diag_l0_me: missing an iseterr
- diag_l0_dumb : fastinit and slowinit now produce errors