scantool_test : Use separate indices for inner and outer loop
cmd_test_readiness has an outer loop to iterate over responses from multiple
ECUs and an inner loop to iterate over the readiness bits in each response,
but the inner loop was reusing the outer loop's index variable! The result
was that the outer loop would only execute once. Fixed.